Consolidate the restart scripts
There were three restart scripts: restart, src/restart, and
src/restart.sh.
All of these did essentially the same thing, and appeared to be the same
script at different points in time.

Replace restart and src/restart.sh with small wrapper scripts that call
src/restart, which will be the official version of the script.

Additionally, adjust src/restart to correctly find the moo binary, no
matter how it's called.

This way, only src/restart needs to be maintained.
The other scripts will act the same way as before, and are only for
compatibility and convenience.