Automatically insert the HellCore version
Since configure.ac already knows the version of HellCore, have the
configure script automatically update the HellCore version in
src/version.c.

When the version of HellCore is changed, it should be instead changed
in the appropriate spot near the top of configure.ac.
The version number can be any arbitrary string.