Add an auto-build script
To simplify things for users, and for the convenience of the developers,
the build.sh script has been added.

The script will automatically determine whether the build system has
been generated (and regenerate it if necessary), configure it (if
necessary), and build the project.

In short, running ``./build.sh`` should perform all the steps necessary
to build HellCore, assuming you have all of the dependencies installed.