Update the README
Update the README for the new autoconf and automake build system.

Add the new (and discovered) dependencies for building HellCore.

Inform the user about how to run the server via the restart script
first, as this is the simpler (and probably preferred) method.

Change the build instructions to tell the user to run the build.sh
script.
Also give information on how to manually run the build system, if they
so prefer.

Additionally, in the markdown README, use the proper formatting to
display the commands to run.
This should both use shell-based highlighting on GitHub, and display it
in monospaced font.