Fix undefined behavior in comparison
Running the restart script through shellcheck revealed a potential issue
with using -a and -o for and/or in shell script tests (SC2166).
Apparently, they're "obsolescent extensions in POSIX and their behavior
is almost always undefined".

Replace this with the safer way of splitting it up into two separate
tests and using the shell's built-in or-logic.