Fix a buffer over-read
When binary_to_raw_bytes() expanded escaped binary sequences of the form
~xy, it didn't check whether x or y were NUL bytes.

Thus, if a string ended in tilde, binary_to_raw_bytes() would count the
NUL byte that should've ended the string as x in the ~xy sequence.

This would cause the next loop to continue reading *after* the NUL byte
that should've signalled the end of the string.

A similar thing would happen if tilde was the second-to-last character
in some situations.

This would cause binary_to_raw_bytes() to output whatever memory came
after the string, potentially leaking information, or, in particularly
unlucky circumstances, potentially causing a segfault.

Fix this by checking if x or y is a NUL byte when processing a ~xy
escape sequence.