Fix the SHA256 implementation when doing 64-bit builds
Currently when building for 64-bit architectures, the SHA256 functions
incorrectly compute checksums in a way that differs from reference
implementations and extant 32-bit implementations.

This is mainly due to relying on certain types to have guaranteed sizes
and overflow behavior, which isn't the case anymore when all longs are
64-bit instead of 32-bit.

In addition, implicitly casting pointers to a different signedness and
constness seemed to scramble them oddly, unless I'm misreading gdb.

Finally, several functions were relying on implicit function
declarations, which assume that an undeclared function returns an int.
These functions return a pointer, which may or may not be the same size
as an int.

Fix all of these issues by using explicitly sized data types and
avoiding odd casts.

This seems to make SHA-256 spit out expected hashes for test cases that
should go through all of the code.

It might still worth looking into using openssl or something as a
reference implementation at some point, to sidestep future issues like
this, and potentially open the door to, e.g., TLS or PBKDF2 and other
newer password algorithms.