Add missing headers
Some files use functions such as strlen() without including the proper
headers.
This causes the compiler to "guess" what the arguments and return type
of the functions are when compiling, which isn't necessarily reliable.

Fix this by including the missing headers to silence warnings and avoid
potential edge cases where the function type may not quite match the
compiler's guess.