Fix an incorrect comparison
db_match_prep() is used to try to look up a preposition in the database.

It does this by first checking if the string is of the form
"[#]<digits>", and looking up the given index in the preposition table.

If the preposition string isn't an index, it instead looks up the string
in the aliases of the preposition table.

The preposition table stores each alias of a preposition separately.
E.g. "with/using" is stored with its alias strings being "with" and
"using", rather than together with the '/'.

db_match_prep() checks if its copy of the string to lookup has a '/' in
it, and replaces it with a NUL, so that if passed, e.g., "with/using",
it would look up "with" in the table, with the '/' and everything after
it cut off.

It does this via strchr(), which returns a pointer to the '/', or NULL
if one wasn't found.

However, this pointer was checked against '\0' instead of NULL.
Fix the comparison by checking the pointer against NULL instead.