Add complex value types to the var_type enum
Currently, complex value types such as TYPE_STR are #defines which just
turn the TYPE_COMPLEX_FLAG bit on in their respective type constant.

This helps some functions identify complex types by checking a single
bit.

However, using these #defines interchangeably with the enum leads to
compiler warnings about invalid values for the var_type enum being used,
and makes it harder for the compiler to "reason" about how the enum is
used.

Fix this by just adding the complex types to the enum with the
appropriate values to allow the compiler to check them.
(This is also the approach toaststunt seems to have taken.)